+++ src/main/java/com/ev/report/service/QualityManagementAccountingReportService.java
@@ -3,6 +3,7 @@ package com.ev.report.service;
 import com.ev.custom.domain.SupplierDO;
 import org.apache.commons.lang3.tuple.Pair;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 /**
@@ -23,7 +24,7 @@ public interface QualityManagementAccountingReportService {
 
     Pair<List<Map<String, Object>>, Map<String, Object>> badPurchaseResult(int showTotalInt, int showItemInt, Long supplierId, Long materielId, String startTime, String endTime);
 
-    Pair<List<Map<String, Object>>, Double> badProcessForDeptResult(int showDeptTotalInt, int showItemInt, int showProcessTotalInt, Long deptId, Long processId, Long operator, Long deviceId, String startTime, String endTime);
+    Pair<List<Map<String, Object>>, BigDecimal> badProcessForDeptResult(int showDeptTotalInt, int showItemInt, int showProcessTotalInt, Long deptId, Long processId, Long operator, Long deviceId, String startTime, String endTime);
 
-    Pair<List<Map<String, Object>>, Double> badProductForDeptResult(int showDeptTotalInt, int showItemInt, int showProductTotalInt, Long deptId, Long materielId, String startTime, String endTime);
+    Pair<List<Map<String, Object>>, BigDecimal> badProductForDeptResult(int showDeptTotalInt, int showItemInt, int showProductTotalInt, Long deptId, Long materielId, String startTime, String endTime);
 }
